requirements.txt
amqp==5.3.1
anyio==3.7.1                 # ← نسخه 4.9.0 وجود ندارد، باعث خطا می‌شود
asgiref==3.8.1
billiard==4.2.1
celery==5.3.6                # نسخه پایدار نهایی
celery-stubs==0.1.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.7                 # آخرین نسخه پایدار کلیک
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
Django==4.2.6
django-cors-headers==4.3.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.3.1
drf-yasg==1.21.10
ecdsa==0.19.1
h11==0.14.0
httpcore==1.0.2
httpx==0.28.1
idna==3.6
inflection==0.5.1
kombu==5.3.4
packaging==23.2
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pyasn1==0.6.1
pycparser==2.22
PyJWT==2.9.0
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
pytz==2024.1
PyYAML==6.0.1
redis==5.0.4
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
vine==5.1.0
wcwidth==0.2.13
pillow
django-jazzmin
